add output info your channel
