+++ src/components/Profile/Profile.tsx
@@ -1,8 +1,34 @@
 import { Outlet } from 'react-router-dom';
-import { PrimaryProcessingButton } from '../common/buttons/PrimaryProcessingButton';
-import { Navbar } from '../common/navbars/Navbar';
+ import { Navbar } from '../common/navbars/Navbar';
+import { useEffect, useState } from 'react';
+import http_api from '../../services/http_api';
+import { useSelector } from 'react-redux';
+import { IAuthUser } from '../../store/reducers/auth/types';
 
 const Profile = () => {
+
+
+  const [userData, setUserData] = useState<IUserInfo>();
+  const { isAuth,user } = useSelector((store: any) => store.auth as IAuthUser);
+  useEffect(() => {
+    fetchData();
+ }, [ setUserData]);
+
+ const fetchData = async () => {
+    try {
+   
+      const response = await http_api.get(`/api/User/GetUser?ChannelId=${user?.userId}`);
+      const data = await response.data
+ 
+      setUserData(data);
+    } catch (error) {
+      console.error('Error fetching user data:', error);
+    }
+ };
+
+ if (!userData) {
+    return <div>Loading...</div>;
+ }
   return (
     <>
       <div className="mt-10 mx-10">
@@ -11,7 +37,7 @@ const Profile = () => {
           <div className="w-36 h-36 dark:bg-primary rounded-full p-3 mr-6">
             <img
               className=" fill-white rounded-full dark:bg-transparent"
-              src="/human.jpg"
+              src= {"/api/Photo/GetPhotoUrl/"+userData.channelPhotoFileId+"/150"}
               alt=""
             />
           </div>
@@ -19,7 +45,7 @@ const Profile = () => {
           <div className="text">
             {/* Channel title */}
             <h1 className="text-white text-3xl mb-2">Your channel</h1>
-            <h3 className="text-white text-xl mb-2">Alisa Konors</h3>
+            <h3 className="text-white text-xl mb-2">{userData.firstName+" "+userData.lastName}</h3>
           </div>
         </div>
         <div className="nav mt-6">
