+++ src/components/Profile/types.ts
@@ -1,0 +1,11 @@
+interface IUserInfo
+{
+  lastName:string,
+  firstName:string,
+  nickname:string,
+  description:string,
+  subsciptions:number,
+  video:number,
+  channelPhotoFileId:string
+
+}
